Adjusted collision of Pick Ups
The Pick Up prefab had a Box Collider that had a scale of 2 instead of 1. That could be an assisting feature to make the game easier for players, but it seems like an oversight. Feel free to revert.

The pickup's scale transform value also had some sort of floating point error (with values like 0.5000001). I rounded those for the sake of tidiness.