Changed the ball to a cube
Replaced the player's mesh and collision with a cube shape, locked the rotation of the cube's rigidbody, and adjusted the player's physics to feel better.

Something important to note is that I disabled gravity for the player to further reduce the friction between the player and the ground. If the level is ever updated to have slopes or jumps that cause the player's Y level to change, we'd need to redo the player's moving and physics.