Updated physics, made everything slippery
Made a new physic material that has zero friction and a bit of bounce, then added that to the player, ground, and walls.

As a result, the player's speed is no longer nullified when touching a wall and gravity could be re-enabled. The player's speed was also increased a bit.